Conflicts:
	src/jdbc_cafe/dto/Cafe.java
	src/jdbc_cafe/ui/CoffeeUi.java

+++ src/jdbc_cafe/dao/CafeDao.java
@@ -64,10 +64,10 @@ public class CafeDao {
 		int cost = rs.getInt(3);
 		int salesnum = rs.getInt(4);
 		int supply = rs.getInt(5);
-		int surtex = rs.getInt(6);
+		double surtex = rs.getInt(6);
 		int price = rs.getInt(7);
 		int margin = rs.getInt(8);
-		int marginprice = rs.getInt(9);
+		double marginprice = rs.getInt(9);
 		
 		return new Cafe(cofcode, cofname, cost, salesnum, supply, surtex, price, margin, marginprice);
 	}
